Use an intermediary cast in is_space
The intermediary cast removes one move to do sign expansion. If this project targeted C#9 we could instead use nint directly.

In addition, I switched the table back to bool. The ROS optimization kicks in for any one byte type. This produces less IL, which improves the chance for this method to get inlined in larger methods.